requirements.txt
numpy
Flask
gunicorn
sklearn
pandas
pyYAML
-e . #(it will installl localy declared packages)